Moved some functions to Util package
